moved mail renderer into scheduler mail channel dir
