+++ src/scheduler/channels/mail/renderer/ResonatorEmail.js
